feat(preset-hugo): support configured post types
