+++ packages/preset-hugo/package.json
@@ -22,6 +22,7 @@
   "main": "index.js",
   "files": [
     "assets",
+    "lib",
     "index.js"
   ],
   "bugs": {
@@ -35,6 +36,7 @@
   "dependencies": {
     "@iarna/toml": "^2.2.5",
     "camelcase-keys": "^9.0.0",
+    "plur": "^5.1.0",
     "yaml": "^2.0.0"
   },
   "publishConfig": {
